move test to external file.
