+++ identity_core/src/lib.rs
@@ -1,5 +1,5 @@
-mod did;
-mod did_parser;
+pub mod did;
+pub mod did_parser;
 mod error;
 
 pub use error::{Error, Result};
